Wire serial/keyboard drivers into BIOS entry points
Replace stub I/O with full IOBYTE-routed dispatch:
- CONOUT: TTY/CRT/BAT/UC1 modes route to SIO-B and/or CRT display
- CONIN: reads from keyboard and/or SIO-B per IOBYTE CON: field,
  applies input conversion table, clears bit 7
- CONST: checks appropriate input source(s) per IOBYTE
- LIST: routes to SIO-B/CRT/SIO-A/parked per IOBYTE LST: field
- PUNCH: routes to SIO-A/SIO-B/parked per IOBYTE PUN: field
- READER: routes to SIO-A (with RTS)/SIO-B/EOF per IOBYTE RDR: field
- LISTST: returns TX ready flag for current LST: device
- Cold boot: DCD auto-detection sets IOBYTE to JOINED or LOCAL

Z80 code+rodata: ~6.6KB (track 0 headroom OK).
9/9 native tests pass, Z80 ticks smoke test exit 0.

+++ src/serial.c
@@ -5,7 +5,7 @@
 #define PIO_A_DATA  0x10
 
 void serial_ch_init(serial_ch_t *ch, byte ctrl_port, byte data_port) {
-    ringbuf_init(&ch->rx_ring, ch->rx_storage, SIO_MASK);
+    ringbuf_init(&ch->rx_ring, ch->rx_storage, (byte)SIO_MASK);
     ch->tx_ready = 0x00;
     ch->wr5_base = 0x60;  /* default: 8-bit TX */
     ch->rr0_cached = 0;
